Refatorar utilitários de processamento e adicionar testes
