+++ processing/__init__.py
@@ -1,0 +1,8 @@
+"""Utilitários para processamento de vídeo do Editor AUTOM TICO."""
+
+__all__ = [
+    "language_utils",
+    "typing_renderer",
+    "ffmpeg_pipeline",
+    "process_manager",
+]
